Updated to debian 11, fixes and macos support

+++ Personalizar/fonts.sh
@@ -26,13 +26,13 @@
 ############################
 ##       FUNCIONES        ##
 ############################
-fuentes_repositorios() {
+fonts_repositories() {
     echo -e "$VE Instalando fuentes desde$RO repositorios$CL"
 
     instalarSoftwareLista "$SOFTLIST/Personalizar/fonts.lst"
 }
 
-fuentes_download() {
+fonts_download() {
     descargar 'Inconsolata Nerd Font Complete.otf' 'https://github.com/ryanoasis/nerd-fonts/blob/master/patched-fonts/Inconsolata/complete/Inconsolata%20Nerd%20Font%20Complete.otf'
 
     descargar 'Inconsolata Nerd Font Complete Mono.otf' 'https://github.com/ryanoasis/nerd-fonts/blob/master/patched-fonts/Inconsolata/complete/Inconsolata%20Nerd%20Font%20Complete%20Mono.otf'
@@ -43,53 +43,29 @@ fuentes_download() {
         sudo mkdir '/usr/share/fonts/Inconsolata'
     fi
 
-    sudo cp "$WORKSCRIPT/tmp/Inconsolata Nerd Font Complete.otf" '/usr/share/fonts/Inconsolata/'
+    if [[ ! -f '/usr/share/fonts/Inconsolata/Complete.otf' ]]; then
+        sudo cp "$WORKSCRIPT/tmp/Inconsolata Nerd Font Complete.otf" '/usr/share/fonts/Inconsolata/'
+    fi
 
-    sudo cp "$WORKSCRIPT/tmp/Inconsolata Nerd Font Complete Mono.otf" '/usr/share/fonts/Inconsolata/'
+    if [[ ! -f '/usr/share/fonts/Inconsolata/Mono.otf' ]]; then
+        sudo cp "$WORKSCRIPT/tmp/Inconsolata Nerd Font Complete Mono.otf" '/usr/share/fonts/Inconsolata/'
+    fi
 
-    sudo chmod 755 -R '/usr/share/fonts/Inconsolata'
+    sudo chmod ugo+r -R '/usr/share/fonts/Inconsolata'
 
     if [[ ! -d '/usr/share/fonts/Powerline' ]]; then
         sudo mkdir '/usr/share/fonts/Powerline'
     fi
 
-    sudo cp "$WORKSCRIPT/tmp/Inconsolata-dz for Powerline.otf" '/usr/share/fonts/Powerline/'
+    if [[ ! -f '/usr/share/fonts/Powerline/Inconsolata-dz for Powerline.otf' ]]; then
+        sudo cp "$WORKSCRIPT/tmp/Inconsolata-dz for Powerline.otf" '/usr/share/fonts/Powerline/'
+    fi
+
 
     sudo chmod 755 -R '/usr/share/fonts/Powerline'
 }
 
-fuentes_nerdfonts() {
-    instaladas='false'
-    if [[ -d "$WORKSCRIPT/tmp/nerd-fonts" ]]; then
-        cd "$WORKSCRIPT/tmp/nerd-fonts" || exit
-        git checkout -- .
-        git pull
-        instaladas='true'
-    else
-        git clone --depth 1 'https://github.com/ryanoasis/nerd-fonts.git' "$WORKSCRIPT/tmp/nerd-fonts"
-    fi
-
-    if [[ $instaladas = 'true' ]]; then
-        echo -e "$VE Las fuentes$RO Nerd-Fonts$VE ya estaban instaladas, reinstalar?$CL"
-        read -p ' s/N → ' SN
-
-        if [[ "$SN" = 'y' ]] ||
-           [[ "$SN" = 'Y' ]] ||
-           [[ "$SN" = 's' ]] ||
-           [[ "$SN" = 'S' ]]
-        then
-            cd "$WORKSCRIPT/tmp/nerd-fonts" || exit
-            ./install.sh
-            cd "$WORKSCRIPT" || exit
-        fi
-    else
-        cd "$WORKSCRIPT/tmp/nerd-fonts" || exit
-        ./install.sh
-        cd "$WORKSCRIPT" || exit
-    fi
-}
-
-fuentes_locales() {
+fonts_locals() {
     for f in $WORKSCRIPT/fonts/*
     do
         if [[ -d "$WORKSCRIPT/fonts/$f" ]]; then
@@ -102,22 +78,11 @@ fuentes_locales() {
 fonts_install() {
     echo -e "$VE Añadiendo$RO fuentes Tipográficas$VE al sistema$CL"
 
-    fuentes_repositorios
-    fuentes_download
-    fuentes_locales
-
-    if [[ -d "$WORKSCRIPT/tmp/nerd-fonts/.git" ]]; then
-        fuentes_nerdfonts
+    if [[ "${DISTRO}" = 'macos' ]]; then
+        fonts_repositories
     else
-        echo -e "$VE ¿Instalar$RO Nerd-Fonts$VE, ocupará más de 1GB (y su descarga)?"
-        read -p ' s/N → ' SN
-
-        if [[ "$SN" = 'y' ]] ||
-           [[ "$SN" = 'Y' ]] ||
-           [[ "$SN" = 's' ]] ||
-           [[ "$SN" = 'S' ]]
-        then
-            fuentes_nerdfonts
-        fi
+        fonts_repositories
+        fonts_download
+        fonts_locals
     fi
 }
